+++ compiler-core/src/type_.rs
@@ -975,6 +975,7 @@ impl ModuleValueConstructor {
 #[derive(Debug, Clone)]
 pub struct ModuleFunction {
     pub package: EcoString,
+    pub deprecation: Deprecation,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -1128,6 +1129,7 @@ impl ModuleInterface {
 
         Ok(ModuleFunction {
             package: self.package.clone(),
+            deprecation: value.deprecation.clone(),
         })
     }